Move distance chart to its own component
